+++ src/tensor_network_viz/einsum_module/__init__.py
@@ -1,3 +1,4 @@
+from ._equation import parse_equation_for_shapes
 from .renderer import plot_einsum_network_2d, plot_einsum_network_3d
 from .trace import EinsumTrace, einsum, pair_tensor
 
@@ -5,6 +6,7 @@ __all__ = [
     "EinsumTrace",
     "einsum",
     "pair_tensor",
+    "parse_equation_for_shapes",
     "plot_einsum_network_2d",
     "plot_einsum_network_3d",
 ]